requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.4
black==25.1.0
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-resized==1.0.3
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
fonttools==4.59.2
gunicorn==23.0.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
kiwisolver==1.4.9
kombu==5.6.1
matplotlib==3.10.6
moviepy==2.2.1
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
proglog==0.1.12
prompt_toolkit==3.0.52
psycopg==3.3.1
psycopg-binary==3.3.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==7.1.0
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.4.0
yt-dlp==2025.9.26
